Add className and children props to control tooltip with the expected inclusion in PlayerControls
